tns: folder cleanup
